Rename SIGs to TAGs (#68)
This commit changes the naming of "special interest groups" to
"technical advisory groups". There is no change to any of the content or
charters, purely a drop-in replacement of the name.  The name "technical
advisory group" is more descriptive as to the role and responsibilities
of the groups. Furthermore, the observability group chose the TAG
wording when it was created. This completes that renaming.

Signed-off-by: Andrew Ross <andrross@amazon.com>